+++ CourselyLabs-front/src/layouts/AppFooter.vue
@@ -1,0 +1,9 @@
+<template>
+  <q-footer>
+    <q-toolbar>
+      <q-toolbar-title>
+        CourselyLabs - Todos los derechos reservados
+      </q-toolbar-title>
+    </q-toolbar>
+  </q-footer>
+</template>